make dataset id int16
prepare for synthetic datasets and math functions

+++ src/sparq_types.hpp
@@ -14,7 +14,7 @@
 
 struct sparq_dataset_t
 {
-    uint8_t id = 0;
+    int16_t id = 0;
     char name_buffer[64] = {0};
     std::string name;
     ImVec4 color = ImVec4(1, 0, 0, 1);
